tokenomy: add a Client interface for v1 and v2
Currently, the Client provide an interface for trading functionalities.

In order for Client to be able to talk with two API versions there is
some changes that need to made on v1.  The return values of v1 need to
match with value of v2, so we move the TradeResponse, Order, and
TradePrice to root package.