rawfloat: refactoring the String format
Previously, if the Rawfloat contains zero digits greater than
MaxPrecision it will still print the result as non-zero.  For example,
0.000_000_001 will be printed as "0.000_000_001".

This commit change the rule so it will print 0, because the value,
when rounded to MaxPrecision will accumulated to 0.

This commit also introduce new method `Round()` to round the Rawfloat
value into n-digit max precision value.