trade: remove the ambiguous TradeID field
When opening a trade, or calling UserTrade, UserTradesOpen,
UserTradesClosed the ID field represent the trade ID in the platform.
But when we call UserTrades, the trade ID is stored in TradeID field
while ID field represent the internal ID in the platform.

To minimize this confusion and to provide more consistency, we remove
the TradeID and use the ID field to represent all of trade ID information.